Former-commit-id: 36d4ef5c56a6abf08ee400e806ce8fdc65e565c0

+++ Script/bundle_app.py
@@ -25,8 +25,9 @@ def main(args):
     app_path = os.path.join(build_path, 'app')
 
     if not args['in_place']:
-        LOGGER.info('=== Removing Old App Directory ===')
-        shutil.rmtree(app_path)
+        if os.path.exists(app_path):
+            LOGGER.info('=== Removing Old App Directory ===')
+            shutil.rmtree(app_path)
 
         if args['serverless']:
             LOGGER.info('=== Copying Dependencies ===')
